Crate riddle_renderer_common

Riddle crate containing providing the common API to which riddle renderers abide. This allows secondary libraries to be defined in terms of the traits and structs defined in this crate without needing to encode knowledge of any specific renderers.

Modules

vertex

Structs

Renderable

SpriteFont
An efficient riddle_font::ImgFont renderer.

SpriteInitArgs

SpriteRenderArgs

Enums

FilterMode

RendererError

Traits

CommonRenderer
The root object of a renderer implementation, associated with a single display.

CommonSprite
Sprites are conceptually both a reference to an image, and the sub region of the image which represents the logical sprite.

RenderContext
The context provided to render callbacks