Struct riddle_renderer_common::SpriteFont

pub struct SpriteFont<R: CommonRenderer> { /* fields omitted */ }

An efficient riddle_font::ImgFont renderer.

Use crate::SpriteRenderArgs for access to all supported paramters when rendering sprites.

Example

let renderer = Renderer::new_from_window(&window)?;

// Load a TTFont, generate an ImgFont, and then construct the SpriteFont with it
let img_font: ImgFont = {
    let font_bytes = include_bytes!("../../example_assets/Roboto-Regular.ttf");
    let ttfont = TTFont::load(&font_bytes[..])?;
    ImgFontGenerator::new("abcdefghijklmnopqrstuvwxyz ", 32).generate(&ttfont)?
};
let sprite_font = SpriteFont::new(&renderer, img_font)?;

// Render the sprite at the top left corner of the screen
renderer.render(|render_ctx| {
    render_ctx.clear(Color::WHITE);
    sprite_font.render(
        render_ctx,
        &SpriteRenderArgs::new([0.0, 0.0]).with_color(Color::BLACK),
        "hello world",
    )
})?;

Implementations

impl<R: CommonRenderer> SpriteFont<R>

pub fn new(renderer: &R, font: ImgFont) -> Result<Self, RendererError>

Build a SpriteFont from the ImgFont given. The ImgFont’s image will get loaded in to a texture, and its glyph information will be used for layout.

pub fn render<Ctx: RenderContext<R> + ?Sized>(
&self,
render_ctx: &mut Ctx,
render_args: &SpriteRenderArgs,
text: &str
) -> Result<(), RendererError>